chore: move integration test and update unit test location
